+++ CHANGELOG.md
@@ -75,6 +75,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Added
 
+- Added new `actor.actions.flash(...)` `Action` to flash a color for a period of time
 - Added a new `ex.NineSlice` `Graphic` for creating arbitrarily resizable rectangular regions, useful for creating UI, backgrounds, and other resizable elements.
   ```typescript
   var nineSlice = new ex.NineSlice({